Label plotting channels
Now, if there is a different number of channels in the loaded sigInfo than in current sigTbl, it will try to find for each channel in sigInfo, to which channel in sigTbl it corresponds and renumber the channels in the lblSet. Then, it will update sigInfo (like it always did).